linear-airtable-sync/pyproject.toml

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "linear-airtable-py"
version = "0.1.0"
description = "Syncs linear to airtable (one-way)"
authors = ["rigélblu <contact@rigelblu.com>"]

[tool.poetry.dependencies]
fastapi = "^0.100.1"
flask = "^2.3.2"
functions-framework = "^3.4.0"
gunicorn = "^20.1.0"
pyairtable = "^1.5.0"
python = "^3.9"
google-cloud-logging = "^3.6.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^23.7.0"
flake8 = "6.0.0"
mypy = "1.4.1"
pre-commit = "3.3.3"
pytest = "^7.2.2"

[tool.black]
line-length = 79
target-version = ['py37', 'py38', 'py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | env
)/
'''
